Internal scale up buffer wip

+++ src/core/memory.h
@@ -12,8 +12,8 @@ inline void ZeroSize(memory_index Size, void* Ptr) {
 }
 
 // Alignment in bytes
-auto inline is_aligned(void* memory, u32 alignment) -> bool {
-  return (((uintptr_t)memory) & (alignment - 1)) == 0;
+auto inline is_aligned(void* memory, u64 alignment) -> bool {
+    return (((uintptr_t)memory) & (alignment - 1)) == 0;
 }
 
 void clear_memory(void* memory, u64 size);
